requirements.txt
pandas==1.4.4
PyInquirer==1.0.3
snowflake-connector-python==2.7.12
snowflake-sqlalchemy==1.4.1
SQLAlchemy==1.4.41